ecourseapi/requirements.txt
asgiref==3.8.0
certifi==2024.2.2
cloudinary==1.39.1
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
